+++ internal/datapath/objects.go
@@ -16,7 +16,7 @@ You should have received a copy of the GNU Affero General Public License
 along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
 
-package gobpf
+package datapath
 
 import (
 	"encoding/binary"
