+++ internal/datapath/objects_internal_test.go
@@ -18,7 +18,7 @@ You should have received a copy of the GNU Affero General Public License
 along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
 
-package gobpf
+package datapath
 
 import (
 	"net"
